chore: rename package
