+++ backend/internal/vendors/kuaikanmanhua/kuaikanmanhua_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package kuaikanmanhua
 
 import (
 	"flag"
