+++ backend/internal/vendors/kuaikanmanhua/vendor_service.go
@@ -1,4 +1,4 @@
-package baozimh
+package kuaikanmanhua
 
 import (
 	"context"
