+++ backend/internal/vendors/kuaikanmanhua/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package kuaikanmanhua
 
 import (
 	"context"
