+++ backend/internal/vendors/manhuagui/manhuagui_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package manhuagui
 
 import (
 	"flag"
