+++ backend/internal/vendors/manhuagui/vendor_service.go
@@ -1,4 +1,4 @@
-package baozimh
+package manhuagui
 
 import (
 	"context"
