+++ backend/internal/vendors/manhuagui/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package manhuagui
 
 import (
 	"context"
