+++ backend/internal/vendors/manhuaren/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package manhuaren
 
 import (
 	"context"
