+++ backend/internal/vendors/qiman6/vendor_service.go
@@ -1,4 +1,4 @@
-package baozimh
+package qiman6
 
 import (
 	"context"
