+++ backend/internal/vendors/qiman6/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package qiman6
 
 import (
 	"context"
