+++ backend/internal/vendors/u17/u17_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package u17
 
 import (
 	"flag"
