+++ backend/internal/vendors/u17/vendor_service.go
@@ -1,4 +1,4 @@
-package baozimh
+package u17
 
 import (
 	"context"
