+++ backend/internal/vendors/u17/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package u17
 
 import (
 	"context"
