+++ backend/internal/vendors/webtoons/vendor_service.go
@@ -1,4 +1,4 @@
-package baozimh
+package webtoons
 
 import (
 	"context"
