+++ backend/internal/vendors/webtoons/vendor_service_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package webtoons
 
 import (
 	"context"
