+++ backend/internal/vendors/webtoons/webtoons_test.go
@@ -1,4 +1,4 @@
-package baozimh
+package webtoons
 
 import (
 	"flag"
